+++ src/classes/babies/familiars/DemonBaby.ts
@@ -8,7 +8,12 @@ import { Callback, onStageWithNaturalDevilRoom } from "isaacscript-common";
 import { mod } from "../../../mod";
 import { Baby } from "../../Baby";
 
-/** Free devil deals. */
+/**
+ * Free devil deals.
+ *
+ * We use both the `POST_PICKUP_INIT` and the `POST_PICKUP_UPDATE` callback so that we can handle
+ * rerolls.
+ */
 export class DemonBaby extends Baby {
   override isValid(): boolean {
     return onStageWithNaturalDevilRoom();
@@ -16,6 +21,15 @@ export class DemonBaby extends Baby {
 
   @Callback(ModCallback.POST_PICKUP_INIT, PickupVariant.COLLECTIBLE)
   postPickupInitCollectible(pickup: EntityPickup): void {
+    this.checkSetPrice(pickup);
+  }
+
+  @Callback(ModCallback.POST_PICKUP_UPDATE, PickupVariant.COLLECTIBLE)
+  postPickupUpdateCollectible(pickup: EntityPickup): void {
+    this.checkSetPrice(pickup);
+  }
+
+  checkSetPrice(pickup: EntityPickup): void {
     const collectible = pickup as EntityPickupCollectible;
 
     if (this.isPricedDevilRoomPoolCollectible(collectible)) {